rafactor to have the game as a separate component
