+++ public/index.html
@@ -28,7 +28,7 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+    <div id="game"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
